+++ extensions/custom-folder/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Custom folder Changelog
 
+## [Feat Image shades ] - 2024-08-05
+- Added optional toggle - shading effect on the image, switching between a filled mask and shaded tones.
+
 ## [Initial Version] - 2023-12-16
+- Custom folder extension